backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic
pydantic-settings
PyMuPDF
openai
python-dotenv
supabase
PyJWT
httpx
